+++ Src/Mojio/Events/HardEvent.cs
@@ -17,7 +17,7 @@ namespace Mojio.Events
         /// <summary>
         /// force
         /// </summary>
-        public float Force { get; set; }
+        public double Force { get; set; }
 
         public HardEvent()
         {
